Reorganize crypto operations into Crypto module
