+++ YubiKit/YubiKit/Crypto/Hash.swift
@@ -1,0 +1,43 @@
+// Copyright Yubico AB
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+import CommonCrypto
+import Foundation
+
+extension Data {
+
+    internal func sha1() -> Data {
+        var digest = [UInt8](repeating: 0, count: Int(CC_SHA1_DIGEST_LENGTH))
+        self.withUnsafeBytes { bytes in
+            _ = CC_SHA1(bytes.baseAddress, UInt32(self.count), &digest)
+        }
+        return Data(digest)
+    }
+
+    internal func sha256() -> Data {
+        var digest = [UInt8](repeating: 0, count: Int(CC_SHA256_DIGEST_LENGTH))
+        self.withUnsafeBytes { bytes in
+            _ = CC_SHA256(bytes.baseAddress, UInt32(self.count), &digest)
+        }
+        return Data(digest)
+    }
+
+    internal func sha512() -> Data {
+        var digest = [UInt8](repeating: 0, count: Int(CC_SHA512_DIGEST_LENGTH))
+        self.withUnsafeBytes { bytes in
+            _ = CC_SHA512(bytes.baseAddress, UInt32(self.count), &digest)
+        }
+        return Data(digest)
+    }
+}
