+++ YubiKit/YubiKit/Utilities/Data+Extensions.swift
@@ -12,159 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import CommonCrypto
 import Foundation
 
 extension Data {
-
-    internal func aescmac(key: Data) throws(EncryptionError) -> Data {
-
-        let constZero = Data([
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        ])
-        let constRb = Data([
-            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x87,
-        ])
-        let blockSize = 16
-        let algorithm = CCAlgorithm(kCCAlgorithmAES128)
-        let iv = Data(constZero)
-
-        let l = try constZero.encrypt(algorithm: algorithm, key: key, iv: iv)
-        var subKey1 = l.shiftedLeftByOne()
-        if (l.bytes[0] & 0x80) != 0 {
-            subKey1 = constRb.xor(with: subKey1)
-        }
-        var subKey2 = subKey1.shiftedLeftByOne()
-        if (subKey1.bytes[0] & 0x80) != 0 {
-            subKey2 = constRb.xor(with: subKey2)
-        }
-
-        let lastBlockIsComplete = self.count % blockSize == 0 && self.count > 0
-
-        let paddedData: Data
-        var lastIv: Data
-        if lastBlockIsComplete {
-            lastIv = subKey1
-            paddedData = self
-        } else {
-            lastIv = subKey2
-            paddedData = self.bitPadded()
-        }
-        let messageSkippingLastBlock = paddedData.subdata(in: 0..<(paddedData.count - blockSize))
-        let lastBlock = paddedData.subdata(in: messageSkippingLastBlock.count..<paddedData.count)
-
-        if messageSkippingLastBlock.count != 0 {
-            // CBC encrypt the message (minus the last block) with a zero IV, and keep only the last block:
-            let encryptedBlock = try messageSkippingLastBlock.encrypt(algorithm: algorithm, key: key, iv: iv).subdata(
-                in: (messageSkippingLastBlock.count - blockSize)..<messageSkippingLastBlock.count
-            )
-            lastIv = lastIv.xor(with: encryptedBlock)
-        }
-
-        return try lastBlock.encrypt(algorithm: algorithm, key: key, iv: lastIv)
-    }
-
-    internal func bitPadded() -> Data {
-        let blockSize = 16
-        var paddedData = self
-        paddedData.append(0x80)
-        let remainder = self.count % blockSize
-        let zeroPadding = remainder == 0 ? blockSize - 1 : blockSize - 1 - remainder
-        return paddedData + Data(count: zeroPadding)
-    }
-
-    internal func encrypt(algorithm: CCAlgorithm, key: Data, iv: Data? = nil) throws(EncryptionError) -> Data {
-        let mode = iv == nil ? CCMode(kCCModeECB) : CCMode(kCCModeCBC)
-        return try cryptOperation(UInt32(kCCEncrypt), algorithm: algorithm, mode: mode, key: key, iv: iv)
-    }
-
-    internal func decrypt(algorithm: CCAlgorithm, key: Data, iv: Data? = nil) throws(EncryptionError) -> Data {
-        let mode = iv == nil ? CCMode(kCCModeECB) : CCMode(kCCModeCBC)
-        return try cryptOperation(UInt32(kCCDecrypt), algorithm: algorithm, mode: mode, key: key, iv: iv)
-    }
-
-    internal func cryptOperation(
-        _ operation: CCOperation,
-        algorithm: CCAlgorithm,
-        mode: CCMode,
-        key: Data,
-        iv: Data?
-    ) throws(EncryptionError) -> Data {
-        guard !key.isEmpty else { throw EncryptionError.missingData }
-
-        let blockSize: Int
-        switch Int(algorithm) {
-        case kCCAlgorithm3DES:
-            blockSize = kCCBlockSize3DES
-        case kCCAlgorithmAES:
-            blockSize = kCCBlockSizeAES128
-        default:
-            throw EncryptionError.unsupportedAlgorithm
-        }
-
-        var outLength: Int = 0
-        let bufferLength = self.count + blockSize
-        var buffer = Data(count: bufferLength)
-        let iv = iv ?? Data()
-
-        let cryptorStatus: CCCryptorStatus = buffer.withUnsafeMutableBytes { bufferBytes in
-            self.withUnsafeBytes { dataBytes in
-                key.withUnsafeBytes { keyBytes in
-                    iv.withUnsafeBytes { ivBytes in
-                        var cryptorRef: CCCryptorRef?
-                        CCCryptorCreateWithMode(
-                            operation,
-                            mode,
-                            algorithm,
-                            CCPadding(ccNoPadding),
-                            iv.count > 0 ? ivBytes.baseAddress : nil,
-                            keyBytes.baseAddress,
-                            key.count,
-                            nil,
-                            0,
-                            0,
-                            0,
-                            &cryptorRef
-                        )
-                        return CCCryptorUpdate(
-                            cryptorRef,
-                            dataBytes.baseAddress,
-                            self.count,
-                            bufferBytes.baseAddress,
-                            bufferLength,
-                            &outLength
-                        )
-                    }
-                }
-            }
-        }
-
-        guard cryptorStatus == kCCSuccess else { throw EncryptionError.cryptorError(cryptorStatus) }
-        return buffer.subdata(in: 0..<outLength)
-    }
-
-    internal func constantTimeCompare(_ other: Data) -> Bool {
-        guard self.count == other.count else { return false }
-        return zip(self, other).reduce(0) { $0 | ($1.0 ^ $1.1) } == 0
-    }
-
-    internal func sha1() -> Data {
-        var digest = [UInt8](repeating: 0, count: Int(CC_SHA1_DIGEST_LENGTH))
-        CC_SHA1(self.bytes, UInt32(self.count), &digest)
-        return Data(digest)
-    }
-
-    internal func sha256() -> Data {
-        var digest = [UInt8](repeating: 0, count: Int(CC_SHA256_DIGEST_LENGTH))
-        CC_SHA256(self.bytes, UInt32(self.count), &digest)
-        return Data(digest)
-    }
-
-    internal func sha512() -> Data {
-        var digest = [UInt8](repeating: 0, count: Int(CC_SHA512_DIGEST_LENGTH))
-        CC_SHA512(self.bytes, UInt32(self.count), &digest)
-        return Data(digest)
-    }
 
     internal var bytes: [UInt8] {
         [UInt8](self)
